# source/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dear Budget API",
        "description": "API Documentation",
        "contact": {
            "email": "rafacla at github.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/",
            "description": "Dear Budget API Server"
        }
    ],
    "paths": {
        "/accounts": {
            "get": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Get list of user accounts",
                "operationId": "App\\Http\\Controllers\\api\\AccountController::index",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Create a new account",
                "operationId": "App\\Http\\Controllers\\api\\AccountController::store",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "role"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "checkingAccount",
                                            "walletCash",
                                            "investmentAccount",
                                            "creditCard",
                                            "expenseAccount",
                                            "incomeAccount"
                                        ]
                                    },
                                    "initialBalanceDate": {
                                        "type": "date"
                                    },
                                    "initialBalanceAmount": {
                                        "type": "number"
                                    },
                                    "statementClosingDay": {
                                        "type": "number"
                                    },
                                    "statementDueDay": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/accounts/{accountID}": {
            "get": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Get a specific account",
                "operationId": "App\\Http\\Controllers\\api\\AccountController::show",
                "parameters": [
                    {
                        "name": "accountID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Update an account",
                "operationId": "App\\Http\\Controllers\\api\\AccountController::update",
                "parameters": [
                    {
                        "name": "accountID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "checkingAccount",
                                            "walletCash",
                                            "investmentAccount",
                                            "creditCard",
                                            "expenseAccount",
                                            "incomeAccount"
                                        ]
                                    },
                                    "initialBalanceDate": {
                                        "type": "date"
                                    },
                                    "initialBalanceAmount": {
                                        "type": "number"
                                    },
                                    "statementClosingDay": {
                                        "type": "number"
                                    },
                                    "statementDueDay": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Delete a specific account",
                "operationId": "App\\Http\\Controllers\\api\\AccountController::destroy",
                "parameters": [
                    {
                        "name": "accountID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/categories/{categoryID}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get a specific category",
                "operationId": "App\\Http\\Controllers\\api\\CategoryController::show",
                "parameters": [
                    {
                        "name": "categoryID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Update a category",
                "operationId": "App\\Http\\Controllers\\api\\CategoryController::update",
                "parameters": [
                    {
                        "name": "categoryID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "expense": {
                                        "type": "bool"
                                    },
                                    "order": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Delete a specific item",
                "operationId": "App\\Http\\Controllers\\api\\CategoryController::destroy",
                "parameters": [
                    {
                        "name": "categoryID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/subcategories/{subcategoryID}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get a specific subcategory",
                "operationId": "App\\Http\\Controllers\\api\\SubcategoryController::show",
                "parameters": [
                    {
                        "name": "subcategoryID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Update a subcategory",
                "operationId": "App\\Http\\Controllers\\api\\SubcategoryController::update",
                "parameters": [
                    {
                        "name": "subcategoryID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "category_id": {
                                        "type": "number"
                                    },
                                    "order": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Delete a specific item",
                "operationId": "App\\Http\\Controllers\\api\\SubcategoryController::destroy",
                "parameters": [
                    {
                        "name": "subcategoryID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}